+++ markitup/templatetags/markitup_tags.py
@@ -28,12 +28,6 @@ def markitup_media(no_jquery=False):
     include_jquery = not bool(no_jquery)
     return dict(register._markitup_context, include_jquery=include_jquery)
 
-# old deprecated version
-@register.inclusion_tag('markitup/include_all.html')
-def markitup_head(no_jquery=False):
-    warnings.warn(DeprecationWarning('markitup_head template tag is deprecated, use markitup_media instead'))
-    return markitup_media(no_jquery)
-
 @register.inclusion_tag('markitup/include_js.html')
 def markitup_js(no_jquery=False):
     include_jquery = not bool(no_jquery)